research_agent_adk/pyproject.toml
[project]
name = "research-agent-adk"
version = "0.1.0"
description = "ADK Research Agent for enterprise agentic AI research with FastAPI and Cloud Run deployment."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "ADK Samples", email = "noreply@example.com" }]

dependencies = [
  "fastapi>=0.112.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "google-adk",  # ADK toolkit per docs
  "google-cloud-aiplatform[adk,agent_engines]>=1.67.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
  "pytest-asyncio>=0.23.0",
]

[project.scripts]
research-agent-adk = "research_agent_adk.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
index-url = "https://pypi.org/simple"

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]